incorporating feedback
Updated with feedback received so far. I think it's important that we comment on this, amend it, and merge it with master asap so people don't end up commenting on outdated versions. Please also note we need to have a proper new draft before Monday 19th Oct as this is when we agreed we will present to the rest of FAC so they have a week to comment on it before it is voted on (or just accepted) at the biz meeting on Monday 26th Oct. (Correct me if I'm wrong on that point)

@msmichellegar @minaorangina I have tried to address the issue of tone, and specifically have changed the bits you pulled out in #20. I have tried to walk the line between being welcoming, but also being declarative, instructional, and objective as Mina suggested in comments on #18.  

Re CoC - I have added this into the rules but I think for now we should revert to the vaguer language of being welcoming as ~~I think~~ I can guarantee there will be much healthy debate before we can agree on a full CoC  

@des-des I have changed that point based on your comment  

@sofer I added something in about London Living Wage which is basically me deciding policy on the fly (but I think it tracks with previous policies) - do take a look.  

-----

For my money, I don't think this should be softened at all. If it were up to me, I'd have it printed, signed, then laminated. I'd turn it into a FACocratic oath that all new students should recite every morning. I think that amid all the avocados and clapping, we need to make it aggressively clear that this is a business that quite literally cannot afford passengers. Nonetheless I have toned it down as much as I can, please do let me know what you think.